@@ -1,13 +1,13 @@
 package com.nrouge.sudoku.util;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class PossibilitesUtilsTest {
+class PossibilitesUtilsTest {
 
 	@Test
-	public void testGetPossibilitesMasques() {
+	void getPossibilitesMasques() {
 		long pos = 1 + 2 + 4 + 8 + 16 + 32;
 		System.out.println("pos=" + pos + ":" + Long.toBinaryString(pos));
 		long[] t = PossibilitesUtils.getPossibilitesMasques(pos, 5);
