@@ -4,7 +4,7 @@ import com.nrouge.sudoku.model.Case;
 
 /**
  * 
- * @author Nicolas Roug�
+ * @author Nicolas Rougé
  */
 public interface ICaseChangeListener {
 	void caseHasChanged(Case c);
