@@ -9,7 +9,7 @@ import com.nrouge.sudoku.solver.UnsolvableCaseException;
 
 /**
  * 
- * @author Nicolas Roug�
+ * @author Nicolas Rougé
  */
 public class JNISolver implements ISolver {
 
