@@ -9,7 +9,7 @@ import java.io.ObjectOutputStream;
 
 /**
  * 
- * @author Nicolas Roug�
+ * @author Nicolas Rougé
  */
 public final class ContextUtils {
 	
