@@ -18,7 +18,7 @@ import com.nrouge.sudoku.solver.UnsolvableCaseException;
 
 /**
  * 
- * @author Nicolas Roug�
+ * @author Nicolas Rougé
  */
 public final class ResolutionUtils {
 	
@@ -41,7 +41,7 @@ public final class ResolutionUtils {
 		Map casePanelMap = config.getCasePanelMap();
 		SudokuSwingGUI sudokuSwingGUI = config.getSudokuSwingGUI();
 		if (listener != null) {
-			// en mode de r�solution visuel, on fige les donn�es saisies par l'utilisateur ...
+			// en mode de r�solution visuel, on fige les donn�es saisies par l'utilisateur ...
 			for (Iterator it = casePanelMap.keySet().iterator(); it.hasNext(); ) {
 				Case c = (Case) it.next();
 				CasePanel cp = (CasePanel) casePanelMap.get(c);
@@ -51,7 +51,7 @@ public final class ResolutionUtils {
 			// ... et on repaint tout
 			sudokuSwingGUI.repaint();
 		}
-		// on lance la r�solution
+		// on lance la r�solution
 		ISolver solver = SolverFactory.createDefaultSolver();
 		try {
 			solver.solve(config.getGrille(), config.getResolutionLevel(), listener);
@@ -61,7 +61,7 @@ public final class ResolutionUtils {
 			JOptionPane.showMessageDialog(sudokuSwingGUI, mse.getMessage(), "Erreur", JOptionPane.ERROR_MESSAGE);
 		} catch (UndeterminedSolutionException e) {
 		}
-		//on fige les classes r�solues et on resauvegarde les possibilit�s
+		//on fige les classes r�solues et on resauvegarde les possibilit�s
 		for (Iterator it = casePanelMap.keySet().iterator(); it.hasNext(); ) {
 			Case c = (Case) it.next();
 			CasePanel cp = (CasePanel) casePanelMap.get(c);
@@ -78,10 +78,10 @@ public final class ResolutionUtils {
 			this.config = config;
 		}
 		public void run() {
-			JDialog dialog = new JDialog(config.getSudokuSwingGUI(), "R�solution", true);
+			JDialog dialog = new JDialog(config.getSudokuSwingGUI(), "R�solution", true);
 			dialog.setDefaultCloseOperation(JDialog.DO_NOTHING_ON_CLOSE);
 			JPanel panel = new JPanel();
-			panel.add(new JLabel("R�solution en cours..."));
+			panel.add(new JLabel("R�solution en cours..."));
 			dialog.setContentPane(panel);
 			dialog.pack();
 			SolverThread st = new SolverThread(config, dialog);
