@@ -9,25 +9,19 @@ import lombok.Setter;
 
 /**
  * 
- * @author Nicolas Roug�
+ * @author Nicolas Roug�
  */
 @Getter
 @Setter
 public class SwingGUIConfig extends GUIConfig {
-	
-	/**
-	 * serialVersionUID
-	 */
-	private static final long serialVersionUID = -509552238453735138L;
-	
-	
+
 	private static final float VALEUR_FONT_SIZE = 24f;
 	private static final int VALEUR_FONT_STYLE = Font.BOLD;
 	private static final float POSSIBILITE_FONT_SIZE = 10f;
 	private static final int POSSIBILITE_FONT_STYLE = Font.PLAIN;
 
 	/**
-	 * R�f�rence sur la frame principale
+	 * R�f�rence sur la frame principale
 	 */
 	private transient SudokuSwingGUI sudokuSwingGUI;
 	
@@ -37,17 +31,17 @@ public class SwingGUIConfig extends GUIConfig {
 	private final Font valeurFont;
 	
 	/**
-	 * Fonte pour l'affichage des possibilit�s
+	 * Fonte pour l'affichage des possibilit�s
 	 */
 	private final Font possibiliteFont;
 	
 	/**
-	 * Mappe de correspondance entre les cases et leur �quivalent graphique, CasePanel
+	 * Mappe de correspondance entre les cases et leur �quivalent graphique, CasePanel
 	 */
 	private Map casePanelMap;
 	
 	/**
-	 * Indique si la r�solution doit �tre visible par l'utilisateur
+	 * Indique si la r�solution doit �tre visible par l'utilisateur
 	 */
 	private boolean visualResolution;
 
